emulation_system/pyproject.toml
[tool.poetry]
name = "emulation-system"
version = "2.3.2"
description = "Repository for connecting all hardware emulations into a single cohesive system"
license = "Apache-2.0"
repository = "https://github.com/Opentrons/opentrons-emulation"
authors = ["Opentrons <engineering@opentrons.com>"]

[tool.poetry.dependencies]
python = "~3.10"
rich = "^12.5.1"

[tool.poetry.dev-dependencies]
pytest = "7.0.1"
pytest-cov = "2.10.1"
pytest-xdist = "~=2.2.1"
pytest-lazy-fixture = "0.6.3"
pydantic = "1.10.2"
decoy = "~=1.9.0"
mypy = "0.981"
flake8 = "~=4.0.1"
flake8-annotations = "~=2.7.0"
flake8-docstrings = "~=1.6.0"
flake8-noqa = "~=1.2.1"
black = "22.3.0"
pyyaml = "6.0"
isort = "5.10.1"
types-pyyaml = "6.0.4"
mdformat-gfm = "^0.3.5"
mdformat_frontmatter = "^0.4.1"
mdformat_footnote = "^0.1.1"

[tool.pytest.ini_options]
addopts = "--color=yes"

[tool.mypy]
plugins = ["pydantic.mypy", "decoy.mypy"]
ignore_missing_imports = false
check_untyped_defs = true
show_error_codes = true
exclude = [
    "emulation_system/compose_file_creator/output/compose_file_model.py"
]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[tool.isort]
profile = "black"

[tool.black]
# LEAVE THE LINE LENGTH AT 240!!!
# make format uses a command line option to override this to 88
# It has to be at 240 for the datamodel-code-generator in generate_compose_spec.sh to not wrap long lines
# datamodel-code-generator uses black under the hood and it will look at this file to figure out what to set line-length at
line-length = 240


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
